NPE: Attempt to get length of null array at ConnectionServiceWrapper.createConnection
[Rootcause]
The latest AOSP change in ArrayMap.java caused a system server
crash if the extra bundle is immuatable empty.  In phone app, when it needs
add additional extras to the bundle passed by the call, it only calls
extra.clone() in an if() block.  If that if() block is not executed, a system
crash will happen.

[Solution]
Always call extra.clone() to avoid the crash.

Bug: 421051442
Bug: 421990909
Test: atest ConnectionServiceWrapperTest.java
Test: atest CreateConnectionProcessorTest.java
Flag: EXEMPT bugfix

Change-Id: Ifc6ca3f0d3d2e8e844f57e84615c2b75176d0664