Mark the call as Voice if it is not VoLTE call.
The carrier doesn't support emergency IMS call and it expects no
VoLTE icon for emergency calls. But the emergency number is truely
called via IMS during dailing. The current AOSP design will mark
the call as VoLTE in call log if it's used to be a VoLTE call.

Bug: 419135856
Test: Manullay Verified
Flag: NONE UX only fix.

Change-Id: I292eb3342130c6f30a645ada38eeeaef2617ebb3