Fix memory leak issue of VideoProviderProxy
[Detail]
TeleService will make a binder and send the binder to VideoProviderProxy
when make a volte call, and then VideoProviderProxy will register
DeathRecipient event which will be received when the binder died,
but as the current design, TeleService is in phone process and the
process is never be killed, so VideoCallImpl will never receive the
event of DeathRecipient and the DeathRecipient object cant be freed.

[Solution]
As VideoProviderProxy does nothing when receive the DeathRecipient event,
there is no need to register the event.

Flag: EXEMPT BUGFIX
Bug:428620859
Test: make
Test: atest TelecomUnitTests

Change-Id: I57720919b78a27b5ecd08e927e4969f1e4970b55