Add support for HD+ indicator in call log.
[Description]
As of today, Dialer shows the HD+ icon in the Incall UI only,
when EVS codec is used.

The purpose of this feature is to extend the existing support
to also show the HD+ icon also in the call history.

[Changes]
Add a new method to check if the call audio quality is
considered as HD+ and when it is, set the HD_PLUS_CALL
flag to call features.

Bug: 428115434
Flag: com.android.server.telecom.flags.hd_plus_call
Test: CTS verified
Change-Id: I3844eb74a88322c860ef2a1e71b3eb34e85fed95