Resolve dependency on XmlUtils.nextElementWithin.
This is a trivial method; just copying it into PhoneAccountRegistrar to
remove the dependency.  Given the trivial nature of the test, I just set
the unit tests to use the flagged path (it did run fine with the unflagged
path as well).

Flag: com.android.server.telecom.flags.resolve_hidden_dependencies_two
Test: Run Telecom unit tests.
Test: Run Telecom CTS tests.
Fixes: 308471542
Change-Id: I1b961ed3e1aef9e90bf6d7e308a811bb71fe0fec