Cleanup default on speaker on wired headset disconnect
Upon investigation, there was a different issue with call audio routing
to speaker when a wired headset was connected and the user turned on/off
speaker. After the headset was disconnected, audio was routing to
speaker because originally we were looking at the original audio route
from the pending audio route. When speaker is turned off, the original
route is still speaker while the destination route is wired headset so
that solution didn't properly address the case when the speaker was
turned off while a headset was plugged in.

This CL modifies this solution further by looking at the user switches
and using that information to determine whether we should default to
speaker when a wired headset is unplugged.

This CL also cleans up the existing flag which has already rolled out
into 25Q2 and is going in as an unflagged change this time.

Bug: 404681140
Flag: EXEMPT bugfix
Test: atest CallAudioRouteControllerTest
Test: Manual test to ensure that when a wired headset is plugged in and
speaker is toggled on/off that we don't route back to speaker when the
headset is unpluffed.
Test: Manual test to ensure that existing cases of speaker turning on
when wired headset is unplugged and speaker was the previous route still
works.

Change-Id: I8ec06592da09db4b94068adaa96f20c5d22fb275