Don't wait on live call disconnect for single sim
Remove the logic in CallSequencingControlelr#makeRoomForOutgoingEmergencyCall to wait for the live call to disconnect in the single sim case where we have an active + held call and we're placing an emergency call. It was found that if the two ongoing calls are merging and we send a disconnect, that it's possible for this transaction to timeout and we end up failing to place the emergency call as a result.

Bug: 394977939
Flag: EXEMPT bugfix
Test: atest CtsTelecomCujTestCases
Change-Id: Ie45f68778bbd8b1b199cbb3984aa9ee3666e33e4