Migrate Telecom uses of appops and permission codes to supported APIs.
A number of these are pretty simple; we were calling startOp on a hidden
proto that uses the int versions of app-op ids through a hidden method;
the supported API is identical but uses a string parameter which is an API.

The PermissionChecker class is entirely hidden; I consulted with the
permissions team and found that PermissionManager is the supported system
API equivalent, hence the new permission requirement.

Note the unit test updates are most just to make sure method parameters
are passed as required; the tests never checked permissions; that is
covered in CTS fully.

Fixes: 308011328
Test: Run CTS tests with flag on to validate app-ops check behavior.
Flag: com.android.server.telecom.flags.resolve_hidden_dependencies_two
Change-Id: I53a6a9567672f08673e4c685e49a81a8d670448a