Fix regression related the incoming call notifier in Telecom.
Telecom has an incoming call notification that is shown in some cases
where answering a new incoming call may cause a call in another app to be
disconnected.  The CTS Verifier suite "telecom incoming self-managed
connection" test verifies that when the user answers an incoming call from
that dialog that one of the ongoing calls is dropped and the incoming call
can be answered.

To correct this, added a "request origin" parameter to the "answerCall"
method in CallsManager.  This is propagated to call sequencing and used
to indicate that the origin of the call answer request is Telecom call
disambiguation taking place in the incoming call notifier.  This allows
sequencing to allow a managed call to be disconnected in favor of a
voip call, even though that is against our usual sequencing policy.

Flag: NONE Bug fix.
Test: Manual test by re-running the CTS verifier test.
Test: Added testAnswerCallAcceptedFromTelecom in the CallSequencingTests
suite.
Fixes: 407486249

Change-Id: I02b2fcaf30b5dfa0a65db30adb646756838de945