Preserve speaker on unhold
When the call audio is routed to speaker and the call is held, when the
call is unheld, ensure that the audio routing remains on speakerphone
unless the user had requested to switch it.

We can reuse the existing mWasOnSpeaker phone logic added to handle
preserving speaker on wired headset disconnects for this logic as well.

Bug: 406898224
Flag:com.android.server.telecom.flags.maybe_default_speaker_after_unhold
Test: Manual testing with call on speaker and holding/unholding the call
as well as verifying speaker does not persist when explicitly moving to
a different route while the call is still held.

Change-Id: Ib3c16c068f26a48a2f6a78a61fd9319578a11a82