Complete VerifyCallStateTransaction early on call disconnect
Problem - VerifyCallStateChangeTransaction waits for a specific call state. If the call disconnects unexpectedly before reaching the target state, the transaction might wait for the full timeout. An initial attempt to fix this could lead to IllegalStateException if the future was already completed.

Solution - on call state changed, if the call state is disconnecting or
disconnected, complete the transaction with a failure instead of
waiting for the timeout.

Flag: com.android.server.telecom.flags.cleanup_verify_call_state
Bug: 404957140
Test: new unit test
Change-Id: I6fe9902a32cb97f39837e5ba5f5de8a7cb064640