Fix cross profile available account determination.
When determining which phone accounts can be used for placing an outgoing
call, consider ongoing calls across all profiles.  This corrects an issue
where an active voip call in the work profile prevents the user from
placing a call on the mobile network in their personal profile.

Flag: NONE Bugfix
Test: Manual test across profiles.
Test: Run unit tests and update.
Fixes: 390216890
Change-Id: I5653549a859c3087689fb96664ca34acee18b367