Fix NPE in BT audio connected states.
It's possible that we receive BT_AUDIO_CONNECTED/BT_AUDIO_DISCONNECTED signals from CallAudioRoutePeripheralAdapter. This was mainly being used as part of the legacy audio routing code and isn't required since we listen to the broadcasts in BluetoothStateReceiver already. Add a NPE check to disregard these messages if the device isn't specified.

Bug: 410872585
Flag: EXEMPT bug fix
Test: atest CallAudioRouteControllerTest
Change-Id: Idf50dfbe2200fa1e2de5d0a2c3a01d7561e0d638