Refactor dependency on SystemVibrator.
SystemVibrator is a hidden framework class that extends the base
Vibrator class.  It has an implementation that essentially wraps all
the APIs in VibratorManager.

To refactor this dependency, created a VibratorAdapter in Ringer which
exposes the method signatures used in Ringer.java; we will either wrap
SystemVibrator or the equivalent calls to VibratorManager.

Test: Manual vibration testing.
Test: Re-run existing tests with flag on and off.
Flag: com.android.server.telecom.flags.resolve_hidden_dependencies_two
Bug: 308205987
Change-Id: Ic333df273d02ac4e75dfc9580053ebce58859c3b