Initial call sequencing metrics impl
Add implementation support for the new call sequencing stats atom. This
will include information like what the primary/secondary call types are
as well as if they are emergency calls and if the phone accounts are the
same. There is another atom (CallSequencingOperationStats) being
included here but the Telecom implementation will be included as part of
a separate CL.

Bug: 384949208
Flag: com.android.server.telecom.flags.call_sequencing_metrics
Test: atest CallTest
Test: atest TelecomPulledAtomTest

Change-Id: Ib0f44c9e365c8c0dee135aa391d6a9a8448201ea