Clean up use refactored audio routing flag
This flag is ready for clean up. There are some associated tests which
were written to support the legacy audio routing structure and have been
modified accordingly to support the new structure. Do note that the use
of this class in the new code path is very small but it is still applicable.

Bug: 409607970
Flag: EXEMPT flag cleanup
Test: atest TelecomUnitTests
Change-Id: I9dad3198e4b41fcf54ee2b7fcdbcde95bfb8a7d8