Prevent self managed calls from logging when not indicated by PhoneAccount.
Currently self managed calls are generally not logged to the call log. As part of the call sequencing refactor, a few logCall invocations were added that did not check if the call was self managed before logging it to the call log. This change adds a check before the invocations of logCall.

Test: manual
Bug: 405942024
Flag: com.android.server.telecom.flags.prevent_self_managed_call_logging
Change-Id: Ie215d25c94cd8e1a5635c4d9c4740861a9d2f962