mv cleanupPhoneAccounts(uninstalledPackage) to a registered receiver
As a part of the mainline project for telecom, telecom can no longer
reference @hidden APIs. In this CL, our logic to remove phone accounts
for removed packages depends on a manifest receiver that references
a hidden API.

To mitigate, telecom is moving to a registered receiver that avoids
referencing the hidden API. The core logic remains the same.

Flag: com.android.server.telecom.flags.resolve_hidden_dependencies_two
Fixes: 308213101
Test: unit tests + manual
       (A) remove app with accts on UserHandle=0
       (B) remove app with accts on secondary user
       (C) remove app with accts on work profile
        verified the accts are auto cleaned up by receiver
Change-Id: If5ed66b70446ee4fb03e2796a94a383ae61c1a3f