Check if future is true before timing out
If the BT binding future is completed and we receive
InCallController#onConnected for the BT ICS twice, ensure that we don't
fail the binding and end up disconnecting the BT ICS if the completion
value is true. This only occurs in edge cases where an outgoing call is
placed at the same time an incoming call is created in Telecom. Telecom
allows the outgoing call because the incoming call is not yet in a
ringing call state and we end up trying to bind to BT ICS twice. The
second invocation of InCallController#onConnected was causing the BT
binding future to time out and we unbound from BT ICS as a result.

Bug: 384062290
Flag: EXEMPT bug fix
Test: m Telecom (I've been unsuccessful in reproducing this specific
scenario locally)

Change-Id: I43245a26ad7bdd572671505dfd9c8a8d793c093b