Ensure disconnecting state is notified to listeners
I had originally removed this in the call sequencing code path as I
thought it was a redundant update because we were already sending the
disconnected state onCallStateChanged. This, however, prevented the
intermediate disconnecting state to be notified to the registered
listeners which were causing some test failures.

In the case of a failure, ensure that we reset the disconnecting call
state and notify the listeners as well.

Bug: 417327935
Flag: EXEMPT bug fix
Test: Manual
Change-Id: I9617d4a1924bde04374c6ac29603d5ff168b9686