Omit communication device type change in BSR
When the BT active device switches in the BSR, we currently switch the
communication device type that's stored in PendingAudioRoute with the
new active device. This becomes an issue if an HFP device is connected
and SCO has not yet disconnected before BT stack reports an active
device change. In this case, when Telecom processes the audio routing to
switch to the new active (i.e. LE) device, it will omit from
disconnecting SCO since the communication device type is on LE audio
now. This is causing routing issues because there is a pending SCO
disconnected msg that never ends up getting cleared.

We should just let the AudioRoute class take care of this where we have
a structured means of processing the original route (and disconnecting
SCO) first before we process requesting audio for the destination route.
We can avoid this race condition in this way.

Bug: 417269298
Flag: EXEMPT bug fix
Test: atest CallAudioRouteControllerTest

Change-Id: Ia6264fd271d86aa600e62bc06106d420647ecfb9