Initial support for local voicemail.
Note; this implementation is NOT complete; just breaking it up to make
review a bit more clean.

This CL adds skeletal support for a new type of Telecom service called a
`LocalVoicemailService`.  This provides support for an OEM provided app
that can answer incoming calls automatically, play a greeting and then
record a response.

Key changes:
* AudioModeTracker - monitors the AudioManager audio mode changes; this is
done to ensure the async audio mode change done by CallAudioModeStateMachine
is completed before binding.
* CallState - new telecom states for calls answered for local voicemail
pending confirmation from the connectionservice, and for a call which is
active for local voicemail processing.
* CallAudioManager - track calls in local voicemail state and transition
the CallAudioModeStateMachine to LocalVoicemailFocusState.
* CallAudioModeStateMachine - add new LocalVoicemailFocusState which
changes the call audio mode to mode MODE_CALL_REDIRECT so that the local
voicemail service can access the call's audio.
* LocalVoicemailController - monitors calls; if there is a new ringing
call schedules it for auto-pickup. When the audio mode moves to
MODE_CALL_REDIRECT and there is a local voicemail call, binds to the local
voicemail service.
* CallsManager - new logic for handling answering a call for local voice
mail.
* localvoicemail test app - a bare bones implementation of an answering
machine.

Note: This is a large functionality; it won't be ramped beyond development
until end-to-end tests are added.

Flag: com.android.server.telecom.flags.local_voicemail
Test: Added new test app for local testing; CTS to come.
API-Coverage-Bug: b/405460896
Bug: 394367444
Bug: 405460896
Change-Id: I2f34b676ce511f13f087509bd6ccf653c041e04a