Bump the test timeout for CallAudioRouteControllerTest.
Bumping the timeout from 500ms to 1000ms as we're occasionally hitting
the timeout during test runs on aosp cf. Also change some of the onCallAudioStateChanged invocations to check for at least one invocation as it's possible that we run into a too many invocations error occasionally.

Bug: 413134811
Flag: EXEMPT test fix
Test: atest CallAudioRouteControllerTest
Change-Id: I0821b343165e0c894fe5df01e6111bab9b47f967