Enforce max ringing call check for DSDA only
Update the max ringing call check in processIncomingCallIntent to only
enforce it for calls that require sequencing. For calls on the same
phone account, we'll allow Telephony to handle this in CallManager
instead where it used to be handled in the past. This is causing timing
issues for single sim where the second incoming call is being rejected
in a lab test due to placement of the 2nd call.

Bug: 415680299
Flag: EXEMPT test fix
Test: atest CallsManagerTest
Change-Id: I09b3cc8093fdbd3173c1800b2d6427d4b5648689