Bypass hold in ECC dial for same phone accounts
Ensure that we don't send a hold request in Telecom when the live call
and emergency call phone accounts are the same. We will let this be
handled in Telephony instead. Samsung was reporting an issue with the
ECC not dialing out because Telecom's hold request was interefering with
Telephony's internal handling.

Bug: 419729271
Flag: com.android.server.telecom.flags.bypass_hold_for_ecc_dial
Test: atest CallSequencingTests
Test: Manual test with normal call + ECC
Test: Manual test with backup calling + ECC to verify ECC is placed

Change-Id: I5d9b2eb180e41b35d3aa5742f0ea5338eb96f199