fix abort call not informing clients
A client reported that after a call was aborted, no signal was given
to the application.

Upon investigation, this is due to an edge case where the callback
is not called.

This change ensures all call states will receive the onDisconnected
callback

Flag: com.android.server.telecom.flags.echo_abort_transactional_outgoing
Bug: 354122175
Test: unit
Change-Id: I885d5f70db336931c14c0beaf0c889dee62f4376