Resolve NPE crash when BT cmpt enabled
When receiving a pkg changed intent notifying Telecom that the BT component is enabled, we were seeing fatal crashes as a result of us already having been bounded to the BT services but no ICS being connected. Prevent the NPE by checking if the key exists in the cached user to BT ICS mapping.

Bug: 422826448
Flag: EXEMPT bug fix
Test: m Telecom
Change-Id: Id0b129b50f4377b48e08303fafd41c7ab10874e5