Bind BT service when profile connected
Ensure we bind to the BT service once we receive indication that the
profile services are connected. In QC's case, we found that the
broadcast for the pkg changed intent is not always sent and the cmpt
stays enabled. If the BT ICS crashes and restarts, it is possible that
we do not bind back to the BT ICS.

Also made some structural changes to ensure that operations are done
being under a lock so that we don't run into races where the BT ICS
reference is deleted but we're still considered connected via the
binding connection. This was causing a PTS test failure in BT tests when
preventing a rebind to the BT services. There was also a bug where the
binding connection reference was being removed before we made a decision
to disconnect from the BT service, which was incorrect logic. We should
really only clean up the references once we do unbind from the service.

Bug: 391057670
Flag: EXEMPT bug fix
Test: atest BluetoothDeviceManagerTest
Test: Manual with toggling on/off BT
Test: Manual with placing calls immediately after one another to confirm
that service registers the newly placed call and unbinding properly
occurs after there are no more calls

Change-Id: Icbe213b47f6517338c86962a104036f6e0d221c7