Make ringer vibration logic user aware
For the vibrate first then ring gradually feature, we should ensure that
this is user aware so that it can work across guest/secondary users. We
should query the context as the provided user (current user in this
case).

Bug: 429213608
Flag: com.android.server.telecom.flags.ringer_vibration_user_aware
Test: Manual verification to verify functionality works after applying
patch on secondary user

Change-Id: I4f45f9b12b1ae583ab153d6d0025ee7a215a9b81