Skip pending msg with communication device check
Skip the pending routing messages if, for example, the device we're
setting the communication device for is already the current
communication device (as reported by the audio fwk). Similarly when
clearing a communication device or disconnecting SCO, check to see if
the communication device is on a different device and skip adding this
message (SPEAKER_OFF + BT_AUDIO_DISCONNECTED).

It is possible that when switching between HFP devices for the audio fwk
to report an intermediate communication update to speaker. Telecom
currently will try to route to speaker and disconnect SCO in the
process. With the new changes, we will only update the UI but if audio
fwk eventually reports SCO as the communication device, Telecom needs to
properly handle routing to the BT device.

Bug: 427837546
Bug: 430908031
Flag:
com.android.server.telecom.flags.skip_pending_msg_if_communication_device_set
Test: atest CallAudioRouteControllerTest

Change-Id: I64cf512eb7a61e2bfbaca0df5f0c66c65cff4966