Ensure clear communication device always called at end
Makes sure that we always call AudioManager#clearCommunicationDevice
at the end of the call. It was seen that once a call was disconnected,
BT stack signaled the SCO audio disconnected signal to Telecom before
we finished cleaning up the audio routing. As a result, we skipped
clearing the communication device at the end. This CL reorders the
clearing logic to ensure that the AudioManager API is always invoked.

Bug: 431000937
Flag: EXEMPT bug fix
Test: atest CallAudioRouteControllerTest

Change-Id: Iea0a0989f79e8e00aed786520c04192d3e2a79bb