Add caller display name for VOIP call
For VOIP calls, we will save the caller display name retrieved from the
call attributes if it is defined. This will be saved in the cached_name
field of the call log which will be used to show the display name.

Bug: 431034555
Flag: com.android.server.telecom.flags.integrated_call_logs
Test: Manual
Change-Id: Ia172ab1206acab20fa4134f282d600250e994e2f