Wait 1s for same phone account live call disconnect
For same phone accounts, wait up to 1s for the live call to be
disconnected before we dial out the ECC. Due to potential races, it's
possible for the call to fail over IMS and be tried over CS because
Telephony hasn't yet processed the request to disconnect the live call.
Instead, we will wait 1s to give enough time in Telephony before we try
dialing the call.

Bug: 429986951
Flag: EXEMPT bugfix
Test: CUJ/Manual
Change-Id: I54d137a3cb69e71c573a4a1c875515c3e68cfab2