Switch ecall live disconnect to orTimeout
Instead of using a countdown latch to interrupt the CompletableFuture,
we can use the built-in orTimeout function. We will complete the future
exceptionally if the live call did not disconnect within a 1s and
proceed with dialing the emergency call after.

Bug: 429986951
Flag: com.android.server.telecom.flags.ecc_wait_for_live_call_disconnect
Test: CUJ
Change-Id: I60e4c5b416ec4137e22fefa590a71e4e54f8ed3e