remove PackageManager#*AsUser hidden references
After investigating a crash related to a mainline bug fix, I realized
I only completed 1/2 of task. The first half involved using a context
object with a userhandle to create a context as a specific user.
The final step was dropping the PackageManager#*AsUser references.

This CL removes all the PackageManager#*AsUser references to complete
the removal of the hidden dependency.

Additionally, I moved up a AnomalyReport in PhoneAccountRegistrar
to ensure we are not seeing any crashes when a non-root user
tries to resolve a component.

Flag: com.android.server.telecom.flags.resolve_hidden_dependencies_two
Fixes: 308009617 (root - resolve hidden dependency bug)
Bug: 424123947 (obsolete crash)
Test: existing
Change-Id: I91d00005f9f384250407653735c10380398db5d5