Update flaky clear pending msgs audio route test
Assert that the pending messages were cleared with an explicit timeout.
This will ensure that the verification has ample time for the previous
message in the queue to be completed. There was some flakiness observed
with these assertions.

There was another flaky test with testing that we route to earpiece
after switching out of speaker while a wired headset is connected and
then we disconnect the wired headset. We should wait on the handler to
finish processing messages in the message queue before asserting the
routing. This was causing some flakiness in test runs.

Bug: 431961317
Bug: 432270350
Flag: EXEMPT bug fix
Test: atest CallAudioRouteControllerTest
Change-Id: I65655f6b63dcd466debfe8427f9218560a8f741f