Preserve audio routing on focus switch
Ensure that we do not calculate the baseline routing every time we
receive a ringing/active focus switch. Instead, keep the routing on the
current or pending audio route. This will address scenarios for
answering calls via quick phrases to not override routing out of speaker
as well as when disconnecting a dock. We can now remove the preserve
speaker logic in the baseline routing as this would be resolved from
preserving audio routing in focus switches.

Bug: 419067749
Bug: 410633186
Flag: com.android.server.telecom.flags.preserve_call_audio_routing
Test: atest CallAudioRouteControllerTest
Change-Id: Ic0b7283d9183b5410307adf042ea7d7ba32930fc