Change Settings.Secure references to use non-user specific prototypes.
This CL is a part of a bigger project to resolve all hidden dependencies
in Telecom in order to become a mainline module.

All references to the hidden Settings.Secure APIs have been removed in favor of the SystemApi equivalents.

Flag: com.android.server.telecom.flags.resolve_hidden_dependencies_two
Bug: 308208703
Test: atest TelecomUnitTests
Change-Id: If473da5eabc442d53219ad46fa354ce62faf1a9c