Keep preferred device selection on focus switch
Preserve preferred device selection if it exists on focus switches to
ringing and active focus. The flagged change currently ensures that we
don't switch the routes but this shouldn't override the preferred device
selection that dictates the default audio route for the call.

Bug: 436310113
Test: atest CallAudioRouteControllerTest
Flag: com.android.server.telecom.flags.preserve_call_audio_routing
Change-Id: Ie9f1bca32b50506aea84f116073d00b818b134a3