Bind to non-UI InCallServices in headless mode.
The `bindToServices` method now checks if the device is in "headless"
mode. If so, it will proceed to connect to non-UI InCallServices even if
the UI InCallService connection fails. This ensures that background
services like Bluetooth or companion apps can still function in headless
environments.

Fixes: 437135421
Fixes: 437136174
Test: Added new test to InCallControllerTests.
Flag: NONE This change is being added for use with the initial launch for microXr.
Change-Id: Id0d70676ef37017857eec29d77f0ec2792cde97b