Handle NPE with null previous communication device
It's possible that for inactive audio routing where the current route is SCO but no communication device has been set so the previous communication device is null. This causes a NPE when calling previousCommunicationDevice.getAddress(). Ensure that this is handled properly and we are also checking the equality of addresses between the previous communication device and the current route's address.

Bug: 438537825
Flag: com.android.server.telecom.flags.ignore_bt_broadcast
Test: atest CallAudioRouteControllerTest
Change-Id: I4ab70fba1d91a84b749336f6e80d4e7b32d5f797