Override skip BT disconnect for inactive routing switch
When we're moving to inactive routing, ensure that we disable the avoid
BT disconnect check in CallAudioRouteController#routeTo since this may
inadvertently skip clearing the communication device at the end of the
call. In the case when the origin route (SCO) is already disconnected,
this field would be set true and we would not clear the communication
device at the end of the call. This causes issues with music playback at
the end of the call. The flag guarding these recent changes has rolled
out into nextfood already, so putting in a forward fix to minimize scope
of impact.

Bug: 439211346
Flag: EXEMPT bug fix
Test: atest CallAudioRouteControllerTest
Change-Id: Ic93ea12f633dc55667006c9e6925d9824216924f