Resolve reference to AudioService#setMicrophoneMute.
This mainline cleanup CL resolves the reference to AudioService#setMicrophoneMute
in Telecom. This method is also referenced in CallAudioStateMachine but that reference is in a legacy code path which is being completely removed as part of a call studio routing cleanup.

Test: atest CallAudioRouteControllerTest.
Fixes: 308482779
Flag: com.android.server.telecom.flags.resolve_hidden_dependencies_two
Change-Id: I51628abb76fe5c9b0f12047e65d417c55fd9ef09