Ensure we only get carrier config for SIM phone accounts.
Add a check to `getCarrierConfigForPhoneAccount` to ensure we only check
the carrier config for `CAPABILITY_SIM_SUBSCRIPTION` phone accounts;
some of the CallSequencing CTS tests are testing various scenarios but they
are NOT using `CAPABILITY_SIM_SUBSCRIPTION` accounts.  That means that
sequencing is relying on the "default" carrier config for the test
phone accounts, which is undesirable.

Specifically, this impacts CallSequencingController#shouldHoldForEmergencyCall;
that method defaults to TRUE if the key isn't found which the CTS was
depending on.

This change doesn't change the behavior of telephony use-cases, but makes
sure that we are not subject to OEM changes to the default carrier config
which are part of telephony.

Test: Added new unit tests for the method.
Bug: 440530209
Flag: EXEMPT BUG_FIX
Change-Id: I5fb8a95539c4c8d7a9114a10d34a67ad62212213