Revise handling the communication device updates
It is possible for the current routing to be on BT already but in an inactive state. This leaves the audio routing in a pending state where it waits for the communication device to be updated to signal SCO audio connected. However, if the current routing is the same as the what the new communication device points to, we skip processing the pending messages. Instead, we should check if the previous communication device is different from the new communication device.

Bug: 441828150
Flag: com.android.server.telecom.flags.ignore_bt_broadcast
Test: atest CallAudioRouteControllerTest
Change-Id: Ibf6ad90dcfd203f6dec19810dac1daf3852ccdfe